+++ src/main/java/org/opentripplanner/api/common/RoutingResource.java
@@ -160,7 +160,7 @@ public abstract class RoutingResource {
    * Whether the trip must be wheelchair accessible.
    *
    * @deprecated TODO OTP2 Regression. Not currently working in OTP2. This is not implemented
-   * in Raptor jet.
+   * in Raptor yet.
    */
   @Deprecated
   @QueryParam("wheelchair")