+++ src/main/java/org/opentripplanner/graph_builder/issues/ConflictingBikeTags.java
@@ -1,34 +1,21 @@
 package org.opentripplanner.graph_builder.issues;
 
 import org.opentripplanner.graph_builder.issue.api.DataImportIssue;
+import org.opentripplanner.openstreetmap.model.OSMWithTags;
 
-public class ConflictingBikeTags implements DataImportIssue {
-
-  public static final String FMT =
-    "Conflicting tags bicycle:[yes|designated] and cycleway: " +
-    "dismount on way %s, assuming dismount";
-  public static final String HTMLFMT =
-    "Conflicting tags bicycle:[yes|designated] and cycleway: " +
-    "dismount on way <a href=\"http://www.openstreetmap.org/way/%d\">\"%d\"</a>, assuming dismount";
-
-  final long wayId;
-
-  public ConflictingBikeTags(long wayId) {
-    this.wayId = wayId;
-  }
+public record ConflictingBikeTags(OSMWithTags entity) implements DataImportIssue {
+  private static String FMT =
+    "Conflicting tags bicycle:[yes|designated] and cycleway:dismount, assuming dismount";
+  private static String HTMLFMT =
+    "Conflicting tags bicycle:[yes|designated] and cycleway:dismount on way <a href='%s'>'%s'</a>, assuming dismount";
 
   @Override
   public String getMessage() {
-    return String.format(FMT, wayId);
+    return String.format(FMT);
   }
 
   @Override
   public String getHTMLMessage() {
-    if (wayId > 0) {
-      return String.format(HTMLFMT, wayId, wayId);
-      // If way is lower then 0 it means it is temporary ID and so useless to link to OSM
-    } else {
-      return getMessage();
-    }
+    return String.format(HTMLFMT, entity.getOpenStreetMapLink(), entity.getId());
   }
 }