+++ src/main/java/org/opentripplanner/graph_builder/issues/StopNotLinkedForTransfers.java
@@ -4,18 +4,11 @@ import org.opentripplanner.graph_builder.issue.api.DataImportIssue;
 import org.opentripplanner.street.model.vertex.TransitStopVertex;
 import org.opentripplanner.street.model.vertex.Vertex;
 
-public class StopNotLinkedForTransfers implements DataImportIssue {
+public record StopNotLinkedForTransfers(TransitStopVertex stop) implements DataImportIssue {
+  private static String FMT = "Stop %s not near any other stops; no transfers are possible.";
 
-  public static final String FMT = "Stop %s not near any other stops; no transfers are possible.";
-
-  public static final String HTMLFMT =
+  private static String HTMLFMT =
     "Stop <a href=\"http://www.openstreetmap.org/?mlat=%s&mlon=%s&layers=T\">\"%s (%s)\"</a> not near any other stops; no transfers are possible.";
-
-  final TransitStopVertex stop;
-
-  public StopNotLinkedForTransfers(TransitStopVertex stop) {
-    this.stop = stop;
-  }
 
   @Override
   public String getMessage() {