+++ src/main/java/org/opentripplanner/routing/api/RoutingService.java
@@ -1,7 +1,7 @@
-package org.opentripplanner.routing.api.request;
+package org.opentripplanner.routing.api;
 
+import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.api.request.RouteViaRequest;
-import org.opentripplanner.routing.api.request.preference.RoutingPreferences;
 import org.opentripplanner.routing.api.response.RoutingResponse;
 import org.opentripplanner.routing.api.response.ViaRoutingResponse;