+++ src/test/java/org/opentripplanner/street/model/edge/StreetEdgeCostTest.java
@@ -2,34 +2,19 @@ package org.opentripplanner.street.model.edge;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.opentripplanner.street.model._data.StreetModelForTest.V1;
+import static org.opentripplanner.street.model._data.StreetModelForTest.V2;
 
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
-import org.opentripplanner.routing.algorithm.GraphRoutingTest;
 import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.street.model.StreetTraversalPermission;
-import org.opentripplanner.street.model.vertex.StreetVertex;
 import org.opentripplanner.street.search.request.StreetSearchRequest;
 import org.opentripplanner.street.search.state.State;
 import org.opentripplanner.test.support.VariableSource;
 
-class StreetEdgeCostTest extends GraphRoutingTest {
-
-  StreetVertex V1;
-  StreetVertex V2;
-
-  public StreetEdgeCostTest() {
-    modelOf(
-      new Builder() {
-        @Override
-        public void build() {
-          V1 = intersection("V1", 0.0, 0.0);
-          V2 = intersection("V2", 2.0, 0.0);
-        }
-      }
-    );
-  }
+class StreetEdgeCostTest {
 
   static Stream<Arguments> walkReluctanceCases = Stream.of(
     Arguments.of(0.5, 37),